+++ src/views/favorite/index.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script>
-import ProductCard from "../../components/Card/index.vue";
+import ProductCard from "@/components/ProductCard/index.vue";
 
 export default {
   components: {
